% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BenchmarkConfig.R
\docType{data}
\name{benchmark_configs}
\alias{benchmark_configs}
\title{Dictionary of Configurations}
\format{
[R6::R6Class] object inheriting from [mlr3misc::Dictionary].
}
\description{
A simple [mlr3misc::Dictionary] storing objects of class [BenchmarkConfig].

For a more convenient way to retrieve and construct tasks, see [tsk()]/[tsks()].
}
\section{Methods}{

See [mlr3misc::Dictionary].
}

\keyword{datasets}
