\name{mfsurrogates-package}
\alias{mfsurrogates-package}
\alias{mfsurrogates}
\docType{package}
\title{
\packageTitle{mfsurrogates}
}
\description{
\packageDescription{mfsurrogates}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mfsurrogates}
\packageIndices{mfsurrogates}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{mfsurrogates}

Maintainer: \packageMaintainer{mfsurrogates}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
