% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mfsurrogates-package}
\alias{mfsurrogates}
\alias{mfsurrogates-package}
\title{mfsurrogates: Surrogate benchmarks for HPO problems}
\description{
Surrogate benchmarks for HPO problems
}
\author{
\strong{Maintainer}: Florian Pfisterer \email{pfistererf@googlemail.com} (\href{https://orcid.org/0000-0001-8867-762X}{ORCID})

Authors:
\itemize{
  \item Lennart Schneider \email{lennart.sch@web.de} (\href{https://orcid.org/0000-0003-4152-5308}{ORCID})
}

}
