% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mfsurrogates-package}
\alias{mfsurrogates}
\alias{mfsurrogates-package}
\title{import data.table
import keras
import tensorflow
import mlr3
import mlr3misc
import paradox
import mlr3keras
import reticulate
import forcats
import checkmate}
\description{
More about what it does (maybe more than one line)
}
