% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mfsurrogates-package}
\alias{mfsurrogates}
\alias{mfsurrogates-package}
\title{mfsurrogates: What the package does (short line)}
\description{
More about what it does (maybe more than one line)
}
